+++ trunk/clientgui/sg_ClientStateIndicator.cpp
@@ -262,7 +262,7 @@ bool ClientStateIndicator::Suspended() {
 bool ClientStateIndicator::ProjectUpdateScheduled() {
     CMainDocument* pDoc     = wxGetApp().GetDocument();
     PROJECT* project;
-    for(int i=0; i<pDoc->GetProjectCount(); i++) {
+    for (size_t i=0; i<pDoc->GetProjectCount(); i++) {
         project = pDoc->state.projects[i];
         if ( project->sched_rpc_pending || project->master_url_fetch_pending || project->scheduler_rpc_in_progress ) {
             return true;
